+++ analytics/RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 ##### Analytics
 * ANDDEP-928 Remove Crashlytics dependencies
 * ANDDEP-1048 Fixing wrong docs links and docs structure
+* ANDDEP-1201 Added navigation modules
 ## 0.4.0
 ##### Analytics
 * ANDDEP-294 Improved analytics module